Move CatalogApi to Api/Catalog namespace with the other apis
